Fix type of read_byte function

+++ include/memory.h
@@ -11,7 +11,7 @@ class Memory {
  public:
     virtual void write_byte(uint16_t addr, uint8_t data) = 0;
     virtual void write_word(uint16_t addr, uint16_t data) = 0;
-    virtual uint16_t read_byte(uint16_t addr) = 0;
+    virtual uint8_t read_byte(uint16_t addr) = 0;
     virtual uint16_t read_word(uint16_t addr) = 0;
     virtual uint8_t &ref_byte(uint16_t addr) = 0;
     // If addr is in memory, this tells memory that a write has been done
